+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/expressions/AnsiCastSuiteBase.scala
@@ -25,6 +25,11 @@ import org.apache.spark.sql.internal.SQLConf
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
 
+/**
+ * Test suite base for both [[Cast]] expression with ANSI mode enabled and [[AnsiCast]] expression.
+ * Note: for new test cases that works for both [[Cast]] and [[AnsiCast]], please add them
+ *       in `CastSuiteBase` instead of this file to ensure the test coverage.
+ */
 abstract class AnsiCastSuiteBase extends CastSuiteBase {
 
   private def testIntMaxAndMin(dt: DataType): Unit = {
